LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scope-shadow.tentative-expected.txt


FAIL anchor-scope scopes to the flat tree assert_equals: expected "10px" but got "20px"
